fix: [Test Class Cleanup Failure -> IO if directory path does not exist to begin with

+++ Backend/DiplomaMakerApi.Tests.Integration/DiplomaMakerApiFactory.cs
@@ -107,7 +107,11 @@ namespace DiplomaMakerApi.Tests.Integration
         {
             var testProjectBinRoot = Directory.GetCurrentDirectory();
             var blobDirectoryPath = Path.Combine(testProjectBinRoot, "Blob");
-            Directory.Delete(blobDirectoryPath, true);
+
+            if (Directory.Exists(blobDirectoryPath))
+            {
+                Directory.Delete(blobDirectoryPath, true);
+            }
         }
 
     }
